Wraparound distortion in Opus
This CL solves the wraparound distortion in Opus.

In the Opus decoder-wrapper we are downsampling the signal from 48 kHz to 32 kHz. This is done in two steps, using the following functions from the signal processing library:
WebRtcSpl_Resample48khzTo32khz() and WebRtcSpl_VectorBitShiftW32ToW16

The latter does not have a saturation check, and the signal can suffer from wraparound. I've added saturation control to the function.

BUG=issue1089

Review URL: https://webrtc-codereview.appspot.com/967004