Test launching script
This script is an attempt to move flags and argumetns to tests from the
buildbot configuration to the source tree.
This will make it easier for anyone to modify test execution behavior
and also has the benefit that it's easier to run the tests in a similar
fashion on a developer workstation.

NOTICE: The audio comparison tool will need to be moved to ~/bin when bots are going to switch over to using this script for execution.

TEST=local execution.
BUG=none

Review URL: https://webrtc-codereview.appspot.com/1021006