Remove use_relative_paths from DEPS
This makes it possible for us to migrate to using the bot_update step
on our buildbots. That would mean they'd use a Git checkout, which
brings stability, speed and best of all: re-enables the
DEPS-second-sync capability on our trybots that we've been lacking.

bot_update currently doesn't support the use_relative_paths variable
so the synced deps end up in the wrong path with it enabled.

Since Chromium doesn't use it, and it doesn't pollute our
DEPS file that much, I think we should switch.

NOTICE: Any custom_deps entries for the solution in .gclient have to be
updated to support this change, which includes the entry normally present
for Valgrind binaries. The bots will need to be updated as well at the
same time as landing this.

BUG=3534
TESTED=Verified a local sync works.
R=andrew@webrtc.org, niklas.enbom@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/13349004