Fix build errors in test programs.
MaxSignalType and memcpy were undeclared in some places.
abs() should be used instead of fabs() on integers.

Add support for building on x64, and link with -lm to get
the trig functions.

With these changes, the test programs can be built for native Linux.

BUG=
R=andrew@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/22919004