Fix an issue with comfort noise in ACMGenericCodecWrapper
In some cases it was not possible to set another payload type for CNG
than the default one. This CL fixes this. The problem was also
dependent on whether the comfort noise codec was registered before or
after the speech codec.

A test is implement to expose the bug, registering comfort noise at a
non-default payload type, and both before and after the speech codec.

BUG=4228
R=kwiberg@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/35199004