Initialize memory in I420VideoFrame unittest
Previously, when CreateEmptyFrame was called with a smaller size than before, we would reuse the allocation. Now, we allocate a new tight frame. The CL that made this change is https://webrtc-codereview.appspot.com/42469004/. This exposed an uninitialized memory problem in a I420VideoFrame unittest. This CL fixes that unittest.

R=stefan@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/42619004