Convert AppRTCDemo and AppRTCDemoTest to proper GYP target.
Initial CL for converting AppRTCDemo and AppRTCDemoTest to
the Chromium style of APK targets. This would
make it possible to get rid of all the ugly
bash stuff we currently have.

CL will bump minimum SDK to v14, but this is the requirement to use Chrome tools.

Initial work was done by kjellander@
https://webrtc-codereview.appspot.com/44549005/

R=kjellander@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/43679004